Website/requirements.txt
numpy
pandas
scikit-learn
streamlit
pillow
opencv-python
PyWavelets